Move fetchIcon into icon utility
